added cascading functionality of neighbouring mines and docstrings
added a function slow_type for responsiveness of outputs to give a typing effect.

added reveal_field function to identify the neighbouring fields which are eligible to reveal if there is no mine nearby.
cascaded the effect of reveal_field to the neighbouring fields with the same logic until a mine was reached.

added 2 different lists to maintain which fields have been revealed, and which are yet to be revealed.

added remove_duplicates function to create a unique list by typecasting set to list.

added a game termination condition upon successfully completing the game, the user is prompted "Congratulations".

docstrings for first few functions